Separate files in folders

+++ include/tileset.h
@@ -9,7 +9,7 @@ class Tileset{
 public:
 	SDL_Texture *texture = NULL;
 	std::vector<Tile> tiles;
-	void init(SDL_Renderer *renderer, std::string path, size_t w, size_t h);
+	void init(SDL_Renderer *renderer, std::string path, int w, int h);
 	SDL_Rect getCord(int i);
 	Tile *actualTile;
 };
